create FaissRetriever class and add BM25Retriever import
